Crate polytope

Defines the Polytope<V> data structure and related methods for constructing polytopes.

Structs

Polytope
A data structure that represents a polytope — a flat sided multi-dimensional object where the vertices are connected by edges, edges by faces, faces by cells, etc.

Type Definitions

Element
An element represented as the list of indices of its lower dimensional bounding elements.